Refactor: Consolidate Tactician training pipelines
Integrate standalone Tactician training pipelines into their respective modules. This simplifies the structure and improves code organization.
